backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
requests>=2.31
urllib3>=2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7
torch>=2.0
numpy>=1.24
